move auction-participant to root, add to workspace
