+++ solana/package.json
@@ -11,11 +11,13 @@
   "exports":{
     ".":{
       "import":"./dist/esm/index.js",
-      "require":"./dist/cjs/index.js"
+      "require":"./dist/cjs/index.js",
+      "types":"./dist/cjs/index.d.ts"
     },
     "./*":{
       "import":"./dist/esm/*/index.js",
-      "require":"./dist/cjs/*/index.js"
+      "require":"./dist/cjs/*/index.js",
+      "types":"./dist/cjs/*/index.d.ts"
     }
   },
   "scripts": {
