+++ solver/package.json
@@ -1,0 +1,17 @@
+{
+    "name": "@wormhole-foundation/example-liquidity-layer-solver",
+    "version": "0.0.1",
+    "author": "Wormhole Contributors",
+    "license": "Apache-2.0",
+    "main": "./dist/cjs/index.js",
+    "types":"./dist/cjs/index.d.ts",
+    "module": "./dist/esm/index.js",
+    "scripts": {
+    },
+    "dependencies": {
+      "@wormhole-foundation/sdk":"^0.7.0-beta.3",
+      "@wormhole-foundation/example-liquidity-layer-solana":"0.0.1",
+      "@wormhole-foundation/example-liquidity-layer-evm":"0.0.1",
+      "@solana/spl-token": "^0.4.6"
+    }
+  }
