+++ solver/src/utils/wormscan.ts
