+++ solver/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends":"../tsconfig.json",
+    "compilerOptions": {
+      "module": "ESNext",
+      "moduleResolution": "Bundler",
+      "rootDir": "./src"
+    },
+    "include": ["src"]
+  }
+
